Use Vec for syntax Map/Set
Using Values as the keys for a map or set does not quite make sense as
the span is part of their equality. This means that the same logical
value with different spans can be inserted in to the set/map. Removing
the span from the equality is also dangerous as the span is an important
part of Value in other contexts (e.g. tests).

Instead use raw Vecs and let higher levels convert them to proper data
structures.