Initial error reporting infrastructure
This adds a Reportable trait which hir and syntax errors implement. This
trait provides a description and an optional span. When reporting an
error we do a linear scan of all files, lines and columns to find the
source position from a byte position. This will be too slow for debug
information but should be fine for reporting.

This also extends our EOG error reporting to include the "context" the
EOF occurred in - e.g. a list, map, string, etc.

This is only properly implemented for a small number of errors as a
proof of concept.