Replace .push() loops with .map().collect()
push() in a loop was being used to be able to return Err() if a problem
was encountered in a loop. I hadn't realised that .collect() can turn a
collection of Result<V, E> in to a Result<Vec<V>, E>.

This should be cleaner, more functional and possibly allow for
clever optimisations within the Rust stdlib.