Include last macro invocation in error messages
This tracks the macro invocation in our scope and uses that to enrich
our errors.

This raises some issues with if using scopes is the correct thing for
expanding a macro in a module context. Currently (def)/(import)/(export)
inside a macro won't bind to the invocation scope even if their bound
identifier comes from that scope. This may not be an issue in practice.