Generic support for destructuring
This defines `fn` in terms of destructuring an argument list. This is
similar to how Scheme works where conceptually an argument list is
passed to the procedure and the rest argument is the tail of that list.

This also allows `def` to use the same destructuring logic so it can be
used to destructure lists.

This doesn't support wildcards or rest arguments but that should be easy
to implement on top of this infrastructure.