First pass of only allowing defs in module body
This makes recursive definition a lot less insane to deal with. It's not
very Lisp-y but it's how Haskell, Elm, Rust, Scala, etc work as well.

The unit tests are hacked by wrapping their data in a function called
`main!`. This is intended intended to be the actual name of the entry
point in the distant future.