Support for `...` in apply
This will pass a list of arguments to a function after all of the
specified fixed arguments. This is both the mirror of the list
destructuring syntax and a replacement for `apply`.

Traditional Scheme-style `apply` cannot be have its type represented
accurately with our type system and requires building runtime
trampolines which I'm trying to avoid. We should be able to add an
`apply`-like macro for ergonomics but wouldn't be usable as a function
value.