Allow arbitrary destruc in parameter lists
We can't exactly specify the `(list)` function type using a list
destructure. This closest is:
```
(defn list #{A} ([vs : A] ...) -> (Listof A)
  vs)
```

This will only return uniform list types and lose the specific type of
the input list. Allowing non-list destrucs allows the following exact
definition:
```
(defn list #{A} [vs : A] -> A
  vs)
```