Remove NonFun/Fun type distinction
Llambda can distinguish most types at runtime by deep inspection - e.g.
it can inspect all elements of a list to determine its type. However, it
cannot distinguish the types of functions as it has no method to encode
a function's signature. This means it supports arbitrary type unions but
cannot have multiple functions in the same union.

As a result Llambda had the concept of a "procedure type" of each union
which was enforced to be a singleton. This was attempted to be
formalised by the NonFun/Fun distinction in Risp.

However, Risp doesn't want to do deep type inspection as in can have an
unbounded cost and happen at surprising times. Instead, disallow
multiple instances of the same top-level "type" in a union.

For example, this would be disallowed:
```
(U (List Int) (List Sym))
```

While this is allowed:
```
(List (U Int Sym))
```

This means that `Fun` is no longer special in its requirements. We'll
need to enforce the top-level type restriction in logic instead of in
the compiler's representation of types.