Add expression type annotations
This uses the form `(: expr Ty)` to annotate the type of expressions.
Typed Racket and Llambda used `(ann expr Ty)` while `(:)` was used to
forward annotate bindings.

We don't support forward annotation and we use `[foo : Ty]` for inline
binding annotations. Try to use the `(:)` form as it's shorter and
doesn't require a new primitive.