Add poly_for_str
This allows us to reconstruct a string for a given ty::Poly. The first
user of this is our type erasure errors but our type check phase will
make heavy use of this.

This also uncovers an interesting corner case where we can't express
they type for this:

```
(defn list #{A} [l : A] -> A l)
```

Here the entire list type of the argument is polymorphic but that can't
be built using `(->)` and therefore stringified.