Introduce complex function type constructors
Previously we had two function type forms: normal and infix. They were
completely identical except for the position of the `->`. The normal
form was similar to other type constructors while the infix form was a
bit easier to visually parse.

We can reuse the "normal" form to take exactly two arguments: the
parameter list type and the return type. This allows us to express the
type of this function:

`(defn list #{A} [l : A] -> A l)`

As this:

`(-> A A)`

It's not clear the above signature cleverness is even useful but it
should be possible to express with a type constructor while it's
allowed.