Remove the concept of unnatural types
This was intended to help our type inference engine to throw type errors
instead if inferring union types. However, the current thinking is to
only allow avoid unnatural types in declarations. For example this would
be allowed:

`(if test 'foo 4)`

But this would not:

`(def x (if test 'foo 4))`

The actual unification happens at the `if` so we're not catching the
unnatural type at the correct place. Instead we should have a function
that tests for unnatural types and refuses to infer them for and
unspecified declaration types.

This also simplifies the (already tricky) code significantly.