Allow discernible types to appear in poly unions
Some of our types can always be discerned at runtime such as symbols and
booleans and their literal subtypes. Currently we will cause a poly
conflict if two poly variables bounded by `Symbol` are put in the same
union as our unification system detects they should be merged. This is
to prevent issues around type erasure.

However, as symbols and booleans are fully discernible we can always
build a type check for them at runtime. Add a special
`poly_is_discernible` function to check for this case and allow the
union.