Introduce `(Cons)` type
This replaces lists with an explicit fixed and rest types with three
types:

- `Nil`: the untyped empty list

- `Listof`: a uniform list of zero or more elements of a given type

- `Cons`: a list with a given head and tail. This is analogous to
  `Pairof` in Typed Racket and Llambda

This allows us to represent a rest list using a `(Cons)` with the rest
list type as the `cdr`.

This allows us to express some common list operations directly within
the type system. For example, `(cons)` has the type `(A B -> (Cons A
B))` and `(car)` has `((Cons A Any) -> A)`. This also reflects our
runtime representation so it can more precisely describe values that can
exist at runtime.

This is closely related to the concept of pairs and improper lists in
Scheme. However, improper lists lead to a number of corner cases in
Scheme which we want to avoid. It's remains to be seen how much improper
lists and pairs will be user-accessible; they could be implicitly
disallowed by e.g. ensuring `(cons)` only accepts lists for its `cdr`.

At the same time this replaces the `Vector` type with two types: a fixed
length `(Vector)` and a uniform `(Vectorof)`. Originally `Vector` was
supposed to be a mirror of `List` that can support efficient pushes and
pops. However, no persistent vector can be nearly as efficient as a list
will be and we're roundtripping through `List` for most monadic
operations anyway. It wasn't worth keeping all the sequential type
complexity around just for more precise `Vector` type.