Number free types using IDs
The type checking pass needs a way to reference free types during
inference and substitution. It could theoretically do this internally by
visiting the HIR in a defined order and numbering the free types itself.
However, that seems complex and fragile.

At the same time this reduces the size of PVarId and FreeTyId to u32 to
save space.