Add first class type predicate type
Type predicates theoretically have the type `(Any -> Bool)` at the
moment. The original plan was to track them specially through the type
checker and trace their type flow to support occurrence typing. However,
this results in ugly hacks in the type checking pass as the type
predicate's function type does not fully describe how types flow through
the function.

This adds a new `(Type?)` type just for type predicates. These are
treated as subtypes of `(Any -> Bool)`. However, the type checker can
apply special logic to `(Type?)` to determine the correct boolean result
for the input type and implement occurrence typing. This can be thought
of as a special form of polymorphism on the input type determinon the
input type determining which literal boolean type is producedd.