Add spans to Destruc
For simple variable definitions it's fine to locate destructuring errors
to the define itself. However, this gets too general in a couple of
cases:

1. If there is a problem with a function's parameter destructure the
   best we can do is span the entire `(fn)`.

2. Assigning a function to a variable of the wrong type would span the
   entire function `(def)`.

This gives each destruc its own span for better error reporting.

This might be a bit overboard; we give every scalar in a fixed list its
own span. This is the easiest way to add the spanning without changing
the way destrucs are represented and the extra span information can't
hurt.