Use a more compact representation for pvar IDs
We need to keep track of which pvars a function introduces for type
checking purposes. Previously we were keeping a vector of `PVarId`s
which are `u32`.

We can by tricky knowing that we allocate pvar IDs sequentially. We can
store a simple range of (`u32`, `u32`) which doesn't require any heap
allocations. This also avoids some temporary vectors when lowering
the function to HIR.