Abstract ty::Fun::pvar_ids
Currently we use a `Range<PVarId>` for both `Poly` and `Mono` functions.
However, it's impossible for a `Mono` function to introduce new poly
variables. This makes it possible to create nonsense Mono functions and
also slightly bloats their size.

Fix by introducing a `PVarIds` trait and associated type. For `Poly`
this is `Range<PVarId>` while for `Mono` its a new unit type created for
this purpose. `Mono::PVarId` was initially implemented using `()`
directly but it feels dirty to implement traits for something a
fundamental as `()`.

This is partially motivated by the upcoming "purity variables" which
will make the problem even worse. This pattern should be easily
extendible to them.