Don't build uninhabitable types during intersect
Avoid using `(U)` as the intersection of two disjoint types.  This is
because if it appears it an "output type" such as a function return or
inferred variable it can mislead our type system in to thinking a value
has the type `(U)`. This is by definition impossible and breaks our type
system.

However, this is safe for parameter types as it's an "input type". Using
`(U)` here simply makes the function unapplicable. This is required to
express our top function types. One change with this PR is we won't
preserve the "shape" of a unified parameter list (e.g. number of fixed
args). There is no immediate need for this as the function cannot be
applied anyway.