Only allow list destruc for function param
We were previously allowing any destruc for params which allowed cute
things like this:

`(defn list #{A} [l : A] -> A l)`

However, this allows the function to take essentially any type and it's
hard to determine its precise arity. This is the reason why the "complex
function" `(Fn)` type constructor exists. This was mildly annoying when
writing the lowering implementation but has caused type checking to
become very complex.

This changes functions to only accept list destrucs. The next step is to
reflect this in the type system so all function types have a well
defined arity.