Introduce (... -> ReturnType) syntax for top funs
We want to dismantle the ability for functions to take arbitrary types
and instead restrict them to the "simple" function types that take fixed
and an optional rest arg. However, we still need to support "top"
function types that accept any param list. This is current represented
by `(Fn (U) ReturnType)`. There are a number of reasons for this:

- The type `(... ->! Any)` is the topmost function type. This is what
  appears in `Any`. Because function parameters are contravariant there
  is no set of parameters every function can accept. This means there is
  no parameter list suitable for the top function type.

- Type inference needs a way to "feed back" the wanted return type for
  self executing functions before it knows the types of its arguments.

Currently this is simply an alias to a function returning an empty param
list but it will eventually correspond to a distinct type.