Add first class `TopFun` type
This replaces our usage of `(U)` params to represent the top function.
This is in preparation for splitting out function params from a single
list type into fixed/rest. It will then be impossible to represent the
top function as a normal function type.

This is a bit gross because of how `is_a`/`unify` deal with supertype
relationships. Now that `TopFun` exists there is a third order subtype
of `Any => TopFun => Fun => TyPred`. We need to explicitly call out all
combinations of that relationship in code.

The alternative is to do what Llambda did and recursively call e.g.
`is_a` with our supertypes. This results in less code but is slower at
runtime as need to re-box new types and recurse for impossible matches
(such as `TyPred` isa `Symbol` recursing as `Fun` isa `Symbol`).

For now live with the pain. A fourth order subtype will probably break
this permanently.