Make ty::Fun take a TopFun
ty::Fun is about to have its parameters split in to fixed/rest which
means it constructor will take 6 arguments. This is a bit crazy so make
it take an explicit `TopFun` to make it more clear what's happening.

This also adds `into_ty` and `into_ref` helpers for the `Fun` type which
replace the adhoc `Ty::new_fun`. This is a lot cleaner and can probably
be added to `Ty<S>` itself