Extremely basic type checking/inference
This only does checking of local functions with no external references.
It's very lightly tested.

The basic approach consists of passing type information between
expression nodes in two directions:

- Once an expression is evaluated it returns the type of the expression
  to its parent. This is called forwards type propagation.

- When we evaluate an expression we pass the type the parent expression
  requires. This is backwards type propagation.

In the case of a function the "forwards" propagation is from the
parameters while the "backwards" propagation is from the return type.

The only type information that doesn't pass between direct neighbours
are the types of parameters and variables. If these aren't explicitly
declared then we will intersect the required type of all of its uses. If
the types are disjoint then type inference will fail with an error.