Make Scope responsible for allocating NsIds
This is a bit tricky. Previously we had a global NsId allocation on the
LoweringContext. Whenever we included a module or expanded a macro we
would allocate a fresh NsId. This means every identifier in the context
would have a globally unique namespace ID. However, this is a bit ugly
to pass around and is testing the patience of the borrow checker.

Instead, we can track the namespace ID per scope. This works because we
only insert bindings in to the most recent scope and once that scope is
exited the bindings (and their associated namespaces) are gone forever.
We only need to start counting where our parent scope ended to ensure
there will be no collisions.

The result of this is once lowering is complete there will only be the
root scope with one namespace for each module or macro invocation at the
module level. All the other macro namespaces will be gone.