Use Rc to avoid copying bindings in to child scope
Currently we copy the entire bindings `HashMap` in to each child scope.
I tried for ~4 hours to get the borrow checker to let me use a reference
here but I failed. For now use `Rc` so we can take a reference on our
parent scope. This should hopefully be low overhead except for `Rc`
allocation.