Merge param and list types
There is a lot of complexity and duplication between list and parameter
types which represent (nearly) the same thing but with a completely
different representation:

- `Params` have a known fixed members and an optional rest member type.
  Only params of different arities are disjoint; params with different
  member types are merged in to a single type. This type was introduced
  to make it easier for our type system to handle functions and their
  applications versus the full power of the list type.

- `Listof`/`Cons`/`Nil` allow for lists to become "optional" at any
  point so non-uniform or multiple rest lists can be allowed. They also
  can have a generic tail at any point. Lists with the same arity but
  disjoint types are also disjoint.

The list types are strictly more powerful. However, they're much more
complex to construct and iterate over. There are also a lot of
conversions between `Params`, `List`, and `destruc::List` where `List`
is quite different from the others.

This essentially promotes `Params` to the primary list type and gets rid
of `Listof`/`Cons`/`Nil`.

At the same time this changes `Vector` to use the same member type
disjoint rules as `List` for consistency. This means only vectors of
disjoint lengths are considered disjoint.