Remove poly type conflict errors
The original intent of our unification code was to avoid having to build
complex or slow type checks. The theory was that if every union member
was easy to distinguish from each other and we only allowed tests for
top-level union members we could avoid complex tests entirely. This
didn't work for multiple reasons:

- With polymorphic union members we can't know if they need to be
  unified with another member as we don't know which specific types will
  be instantiated. This means we would try very hard to unify
  polymorphic types but if we couldn't we would raise a poly conflict
  error.

  This error bubbled up through multiple layers of code (e.g. typeck)
  and eventually would be surfaced to the user. This means that users
  would need to be aware of our type erasure rules and which type checks
  we consider "fast".

- The top-level union member rule disallows some trivial type checks.
  For example, neither `((type-predicate 'foo) Symbol)` or
  `(null? (Listof Any))` select a top-level union member. There were
  some hacks in the type predicate code for this but it was failing in
  the general case.

- Knowledge of runtime type checking abilities would be scattered
  throughout the type system code. Both `unify` and `pred` would need to
  know about runtime type checking in order to work. If these fell out of
  sync with the actual runtime checks it would cause correctness issues

This changes lets us assume that any type check can be performed at
runtime. For functions this can be done by type encoding, and for
collections via length or iteration. This will add definite complexity
for our codegen/runtime but simplifies our type system considerably.

However, this keeps one trick from the previous version of the code. The
unification logic will still unify types that are complex/slow to
distinguish at runtime. While we still need to support the ability to
test arbitrary types this should hopefully reduce the occurrence of
those checks at runtime.