Occurrence typing cleanup
Currently occurrence typing is a bit of a special case hack confined to
`visit_cond` and a couple of functions it requires. They return a
different node type with occurrence typing information which is tricky
to deal with and not very general.

Instead this adds a `type_cond` field to our normal node which is
attached if there is an occurrence typing condition. This now also flows
through `(let)` and `(do)` which makes occurrence typing more powerful.
It also simplifies the code and makes it easier to handle `(not)` and
`(do)` in followup PRs