Clean up source loading
We use `io::Read` everywhere but the first thing we do is read the
entire source in to a string. This is more efficient using
`fs::read_to_string` which will preallocate a string of the appropriate
size. This also removes the pushing of a blank space at the end of the
source; reporting EOF seems to work either way.

This also allows us to only read the file once during the `compile-fail`
tests.