Fix inferring free rest list types
Rest lists are tricky as they have different types "inside" and
"outside" of the destruc. For example, in:

```
(fn (x ...))
```

Within the function x has the type `(Listof Any)` while within the
destruc it has the type `Any`.

This changes the free type to always track the member type while the
`VarType` we use for inference has the full list type.