Handle testing if a polymorphic fun is a top fun
This is important as when we apply functions we pass through the top
function type we need. Currently this fails as we only handle
polymorphism against two functions, not a function and a top.

This also changes the polymorphic function test to instantiate through
the ret type instead of the params type. This should be the same thing
and allows it to be consistent with the top function code.