Remove TypeTagged trait
This was used with `ConstructableFrom` to determine the type tag of the
`Header` we pass. However, this restricts us from implementing
`ConstructableFrom` for tagged unions (e.g. `List`).

Remove the trait completely and make `Header` construction the
responsibility of the `ConstructableFrom` implementations. This is a bit
more verbose but also cleans up our trait zoo a bit.