Hack to unify pair with nil
Our runtime representation of lists uses pairs and nil like a
traditional Lisp. This is hidden from our type system but it can leak
back when dealing with the RFI system. For example, `TopPair` is
actually `(U (List) (List Any Any ...))`.

Fix this by adding a check for this when unifying types. This makes it
come back as the more idomatic type.