Simple REPL implementation
This implements a simple REPL loop using a persistent lowering and type
inference context.

This REPL "dialect" from normal Arret in that we can freely mix
expressions and definitions. We also allow redefinition at the top-level
while it's still forbidden in nested scopes. This allows for a more
natural REPL environment. This could possibly be extended to a "script"
version of Arret in the future.

It would be interesting to see if we could manage to JIT REPL functions
as they're defined. This could hopefully reuse any infrastructure we
use for compile-time evaluation of pure RFI functions without the
restriction on purity.