Simplify processing of deferred defs
We previously tried to process all deferred defs at the end of lowering
the program. This was intended to allow mutually recursive modules but
there are far more fundamental issues to fix first. The result was this
was nearly useless.

This caused complications for the REPL as when modules are imported
their deferred defs aren't processed. Referencing any variables (as
opposed to macros) from these modules would crash the type checker.