Split defs in to per-module definitions
There doesn't appear to be a sane way to allow fully recursive module
includes. Macros can fundamentally change the way parsing is performed
so it isn't as simple as e.g. Go/Scala where the top level definitions
can be extracted in isolation as a first pass.

Rust allows recursive modules along with macros but at least in the 2015
edition there's a lot of ergonomics issues - e.g. requiring macros to be
`use`ed lexically before any invocations of the macro. This might not be
tenable in a Lisp-like where macros are more common.

Not supporting recursive modules has a huge benefit of allowing us to
process all the modules in a package concurrently. A module becomes a
natural unit of work for us to spread across cores/threads. The first
step to this is to split our lowered defs in to module chunks. This
still leaves a lot of global vecs/counters around which will need to be
dealt with next.