More efficient string writing
Currently we write EDN strings char-by-char, even for characters that
don't need escaping. This isn't great for buffered `Write` instances but
it can be pathological for unbuffered ones.

Instead we can loop over the string and only "flush" out the unescaped
characters before we print an escape or reach the end of the string.
This greatly reduces the number of times we need to call in to the
`Write`.