Switch mir::Value from Gc to Rc
Unlike our syntax and expression trees which can be consumed in one pass
our `Value`s need to be persisted. For example, the definition of a
global function needs to be kept around indefinitely for future
inlining. This requires a different memory management strategy.

The initial attempt was to use `Cow<Value>` but that was effectively
just noise as calling any `&mut self` function would require calling
`into_owned` to break the immutable reference the `Cow` held. Nearly
every interesting function that used a value was `&mut self`.

Instead try to substitute `Box` for `Rc` in our `Value`s. The exception
is the `Box<[Value]>` for `List`. As we can't collect in to a `Rc<[]>`
directly we need to use a temporary `Vec`. While `Vec` can be converted
"for free" in to a `Box` by dropping the capacity the same is not true
for an `Rc`.

The result of this is cloning `Value` should become very cheap at the
cost of a small memory/counter overhead.