Introduce `consume_expr`
This allows an expression to be "gifted" to MIR instead of passed by
reference.  The main use of this is avoiding a deep clone of
Fun/RustFun/TyPred when encountering them in definitions.

Normal evaluation of functions still need to be done by reference as
the function can be applied multiple times.