Move garbage collection to a visitor pattern
The current garbage collection implementation in `PartialEval` doesn't
properly recurse in to list `Value`s; it simply looks for `Value`s at
the top level. This will get worse once support function captures.

Because `Value`s can contain each other recursively this is a nightmare
to build an `Iterator` for. Instead we can:

1. Instantiate a `Collection` instance that consumes the old heap

2. Call `visit_box` for each GC root

3. Call `into_new_heap` to retrieve the new `Heap`

This should generate approximately the same code as we were already
calling `visit_box` internally. However, this gives us a lot more
flexibility when visiting GC roots.

This also switches our `Value::List` rest argument from an `Rc` to a
`Box`. This matches the fixed arguments and means we don't need to break
the `Rc` on every garbage collection.