Don't unload RFI libraries during compilation
We need to return the `libloading::Library` instances as part of the
lowered HIR or else they can be unloaded in their `Drop`. By the time we
get to inference/MIR the `RustFun` instances can point to unloaded data.

For some reason this works anyway on macOS but it reliably crashes on
Linux.