Allow running from the root of the workspace
Previously we would always expect our stdlib to be in `..`. This worked
for both tests and the driver as it needed to be run from its crate.

It looks like in nightly Cargo that `cargo run` from the root does the
right thing. This is much nicer but it means we need to be more flexible
in determining our path to accommodate running from:

1. Unit and integration tests
2. The driver crate
3. The root of the workspace

This uses a magic `.arret-root` file to mark our root and then searches
the current directory upwards until it's found. The unit/integration
tests can continue hardcoding `..` via the new `test_paths` method.