Handle panics during partial evaluation
Currently during a panic we simply abort the entire compiler. This is
annoying for two reasons:

1. It skips our normal error reporting so we can't show source locations

2. During REPL use the developer needs to restart the REPL to try again

This adds a Task::panic() to do a "controlled panic" that is then caught
in `PartialEvalCtx`. This is then converted to a new MIR error type and
handled like any other reportable.

At the same time this generalises Reportable handling used by the REPL
to the entire compiler. This reduces repetitive code in both the
compiler and our test drivers.

This has a caveat around panic counts across binaries; see the comment
in `task.rs` for more information.