Run integration tests in parallel
This uses Rayon in `compile-fail`/`run-pass` to run each test case in
parallel. This is a very slight win at the moment but presumably these
tests will start to take the majority of the test time.

This is probably a bit "thrash-y" as `run-pass` and `compile-fail` will
typically be run at the same time by Cargo and they each will spawn a
thread per CPU. Once this becomes an actual issue this can be tweaked
further.

This does slightly more work than previous as it creates a
`SourceLoader` per thread. This means common source files like the
stdlib will be re-loaded and UTF-8 verified multiple times. This should
be more than worth the extra parallelism.