Very basic program generation
The adds a new MIR `Builder` which can generate a series of `Op`s. These
`Op`s are then passed to codegen which uses it to build LLVM IR. This
two-step process is to facilitate optimisation and analysis on the MIR
ops in the future.

The old `EvalMode` as been replaced by the presence or absence of a
`Builder`. If there is no `Builder` we try to evaluate everything
immediately as we can't generate code. Even when generating programs
this is used for evaluating defs.

This is currently almost completely broken. The only working program is
`(exit n)` to return a status code. This works on macOS because it the
"exit" in the stdlib shadows the "exit" in the C library so we
accidentally call it. Linux is not amused by these tricks and crashes.
The next step is to support linking to the appropriate Rust libraries.