Make Fun types always polymorphic
The existing type system has a very black and white view of the world:

- `ty::Mono` can never refer to a polymorphic variable anywhere

- `ty::Poly` can freely be polymorphic

The purpose of `ty::Poly` was to act as a type template which could then
be instantiated in to an instance of `ty::Mono`. An example of an
instantiated type would be the type of a runtime value.

While working on `eval_hir` it became apparent that `ty::Mono` was
useless. Even at runtime a given function value can still be polymorphic
over type/purity variables that it introduces. For instance, this is
identity function is polymorphic at runtime:

```
(fn {X} ([input : X]) -> X input)
```

A given monomorphisation of that function would be "fully" monomorphic
but at that level of abstraction we deal with function signatures, not
`ty::Mono`.

Instead make function types (top and normal) always polymorphic. This
generally simplifies the code except in `subst` which gets much hairier.
Hopefully this is an isolated complexity increase that will be
counterbalanced by a saner type system.