Remove shared TVars, PVars and counters
Having a huge global list of tvars/pvars makes it very difficult to do
parallel or incremental compilation. Instead, move the pvars/tvars
inside the function expressions and types that introduce them.

This adds complexity for each feature that needs to walk the
type/expression trees. They now need to track the "live" tvars for the
node they're visiting. This seems to be less ugly than having to pipe
tvars/pvars through everywhere.

As part of this change we also add support for first class polymorphic
function types in the form `(All #{A} A -> A)`. This was mostly done to
make the unit tests not tied so closely to our internal data structures.

This also avoids passing around counters and instead just uses a shared
global variable that's incremented. This removes some low level tests
that were depending on the specific order that IDs were being allocated.
These were mostly at a lower level of abstraction than is useful to test
now that we have `run-pass`.