Partial support for applying function values
It's impossible to transport a function value from compile time to
runtime unless its MIR value is known. To work around this we keep track
of all functions we box and the value they correspond to. This allows us
to properly apply and build their values.

Additionally when we don't have a builder we can just call the entry
point directly. We're either in the REPL or in a constant context where
the type checker has proven it's safe to call the value.

This also fixes generating thunks for Rust functions that don't return.