Switch to static linking
This is consistent with what Go and Rust do. This produces a standalone
binary that can be freely copied from the system that compiled it.

This is a bit hard on the linker when we link multiple Rust libraries as
they'll all contain bits of the Rust stdlib and Arret runtime and it'll
need to be deduped. This ends up working as expected on Linux and macOS.