Introduce alloc planning
Currently we have each op handle its own allocation when its generated.
This has two limitations:

- There is no way to specify if an allocation should come from the heap
  or the stack. Currently the stack is just hardcoded.

- Allocating from the heap is relatively expensive. We want to amortise
  this cost across as many ops as possible.

This adds a `plan_allocs` which generates a series of `AllocAtom`s. Each
`AllocAtom` begins and ends with the heap in a consistent state. This
allows us to allocate all of the heap for an atom up-front.

This still doesn't implement heap allocation. In fact, this all has no
affect on generated code yet. This is laying groundwork for heap
allocating in the future.