Move task knowledge to codegen
Codegen is now completely responsible for escape analysis and planning
allocations. This means that we don't know if a function will need a
task for heap allocations until it reaches codegen. Currently MIR passes
in a `takes_task` flag but it doesn't know enough to set it to the
correct value.

One exception is when creating thunks or other functions that require a
specific ABI. This is now captured with a `CallConv` enum that can be
used to specify if a function *must* have a task or not.

One cute optimisation we can now perform is skipping task creation if
our Arret main doesn't need it. This would only be hit for the most
trivial of programs but it does mean we can technically generate
programs that don't need the runtime at all.