Fix LLVM optimisation
We were building a module pass builder and then immediately discarding
it and building a new, empty one. This effectively disabled module-level
optimisations.

The trickiest bit of this is that we weren't setting a data layout or
target triple. The LLVM docs seem to imply that failing to set a data
layout prevents a bunch of optimisations. However, it appears that
actually it assumes a particular data layout and performs optimisations
based on that (!). The end result of this was the optimiser getting
confused about the size of structs and building invalid code.