Support for building conditions
This is the bare minimum to get `(assert-eq)` and `(assert-ne)` working.
This also required going back to global reg IDs. This is because a phi
happens in two phases:

1. We plan each branch
2. We examine the terminal ABI types of each branch
3. We convert them to a common type

Step #3 requires "re-opening" an already completed branch which means
we need to do some very funky threading of the RegId back through. For
now just use global reg IDs.