Support for = in our type system
This extends our support for type predicates to `=`. For literal types
this behaves similarly to type predicates. However, for non-literals
this this intersects the types if they compare equal and subtracts
literals if they don't.

There is a bit of a hack with type conditions at the moment. They were
designed to have only one affected type. However, in some cases `=` can
impact the type of two variables. We just pick one at the moment but
this is gross.

This changes the behaviour of type inference slightly. If a condition or
predicate is known to be true/false we remove the expression entirely
and replace it with a static boolean value. And arguments are kept in a
`do` for their side effects. The motivation for this is that the infer
pass has different type information from the eval pass. It's possible
for infer to "see" a constant where eval can't and vice versa. Removing
the construct completely at the infer pass ensures that eval doesn't
needlessly build it.