Hack for JITing dependent functions
Right now we only JIT the "entry" function and any dependencies it might
have are not included. Temporarily fix by JITing all of our dependent
functions for each entry point. We're lazily JITing so the unusued ones
won't actually be compiled by LLVM but we're still generating IR.

This is a temporary fix until we redo how built functions are handled.
This likely involves building them lazily so we can deal with mutually
recursive functions.