Use boxed::Any as the function closure type
We have a number of restrictions on our closure representation based on
our future plans:

1. Closures should be directly reusable when thunking between different
   ABI types. Llambda currently creates a "wrapper" record that points
   to the original function. This is wasteful in both memory and code.

2. We want an unboxed function representation for dealing with typed
   functions. This won't be a proper FunThunk but it still needs to be
   able to point to a closure. This means we should be able to tell if
   the closure is a GC value or not by just looking at the pointer.
   Right now the easiest way to do this is to make it always boxed.

Use a `boxed::Any` so this can point to any GCed value. Initially this
will only handle a single captured value. In the future we can generate
a synthetic record type if we need to capture multiple values